% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sauberung.R
\name{convertDateAndTime}
\alias{convertDateAndTime}
\title{Datum + Uhrzeit konvertieren}
\usage{
convertDateAndTime(x)
}
\arguments{
\item{x}{entsprechender Datensatz}
}
\description{
Datum + Uhrzeit konvertieren
}
