% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sauberung.R
\name{filterStrings}
\alias{filterStrings}
\title{Spalten mit "echten" Strings aussortieren}
\usage{
filterStrings(x)
}
\arguments{
\item{\code{x}}{entsprechender Datensatz}
}
\value{
bereinigter Datensatz
}
\description{
Damit keine Notizen etc fälschlicherweise als String-Variablen aufgenommen werden (eig sind Strings nur als Factors brauchbar)
}
\examples{
df <- filterStrings(dataframe)
}
