% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sauberung.R
\name{resetFactors}
\alias{resetFactors}
\title{"echte" Faktorvariablen auslassen und restliche in Strings umwandeln}
\usage{
resetFactors(x)
}
\description{
"echte" Faktorvariablen auslassen und restliche in Strings umwandeln
}
